Render Login page and fecth User Data
