+++ package.json
@@ -3,20 +3,28 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@fortawesome/fontawesome-svg-core": "^6.4.2",
+    "@fortawesome/free-brands-svg-icons": "^6.4.2",
+    "@fortawesome/free-regular-svg-icons": "^6.4.2",
+    "@fortawesome/free-solid-svg-icons": "^6.4.2",
+    "@fortawesome/react-fontawesome": "^0.2.0",
     "@reduxjs/toolkit": "^1.9.5",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.5.0",
     "bootstrap": "^5.3.2",
+    "npm": "^10.1.0",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
-    "react-axios": "^2.0.6",
     "react-bootstrap": "^2.8.0",
     "react-dom": "^18.2.0",
+    "react-fontawesome": "^1.7.1",
+    "react-icons": "^4.11.0",
     "react-redux": "^8.1.2",
     "react-router-dom": "^6.16.0",
     "react-scripts": "5.0.1",
     "react-test-renderer": "^18.2.0",
+    "react-toastify": "^9.1.3",
     "redux-logger": "^3.0.6",
     "web-vitals": "^2.1.4"
   },
