--- src/app/store.js
@@ -1,8 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit';
-import motorcycleReducer from '../components/motorcycle/motorcycleSlice';
-
-export default configureStore({
-  reducer: {
-    motorcycle: motorcycleReducer,
-  },
-});
