+++ src/components/login/Login-page.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import Login from './Login';
+
+const LoginPage = () => (
+  <div>
+    <Login />
+  </div>
+);
+
+export default LoginPage;
