+++ src/index.css
@@ -1,5 +1,7 @@
 body {
   margin: 0;
+  box-sizing: border-box;
+  padding: 0;
   font-family:
     -apple-system,
     BlinkMacSystemFont,
